+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/lang/XpmVendorType.kt
@@ -31,4 +31,6 @@ interface XpmVendorType {
     fun isValidInstallDir(installDir: String): Boolean
 
     val modulePath: String?
+
+    fun schemaFiles(path: String): Sequence<XpmSchemaFile>
 }