+++ src/plugin-basex/main/uk/co/reecedunn/intellij/plugin/basex/lang/BaseX.kt
@@ -17,6 +17,7 @@ package uk.co.reecedunn.intellij.plugin.basex.lang
 
 import uk.co.reecedunn.intellij.plugin.xpm.lang.XpmProductType
 import uk.co.reecedunn.intellij.plugin.xpm.lang.XpmProductVersion
+import uk.co.reecedunn.intellij.plugin.xpm.lang.XpmSchemaFile
 import uk.co.reecedunn.intellij.plugin.xpm.lang.XpmVendorType
 import java.io.File
 
@@ -41,6 +42,8 @@ object BaseX : XpmVendorType, XpmProductType {
     }
 
     override val modulePath: String? = null
+
+    override fun schemaFiles(path: String): Sequence<XpmSchemaFile> = sequenceOf()
 
     // endregion
     // region Language Versions